The names of the TALEs in a tales object

Each array in a tales object is one TALE, identified by its array_id. This returns those identifiers once each, in the order the arrays appear in x. They are also the names of the vector that tales_rvd_strings returns. A tales_msa has the same arrays as the tales object it was aligned from.

Usage

tales_names(x)

Arguments

x

A tales object.

Value

A character vector of array identifiers.

Details

names() is left to its usual meaning: a tales object is a data frame, so names(x) returns its column names.

See also

Examples

rvd_fasta <- system.file("extdata", "TalA_RVDSeqs_AnnoTALE.fasta",
                         package = "tantale")
x <- as_tales(rvd_fasta, sep = "-")
tales_names(x)
#>  [1] "TalA_BAI3"     "TalA_CFBP1947" "TalA_MAI1"     "TalA_MAI106"  
#>  [5] "TalA_MAI129"   "TalA_MAI134"   "TalA_MAI145"   "TalA_MAI68"   
#>  [9] "TalA_MAI73"    "TalA_MAI95"    "TalA_MAI99"   
length(tales_names(x)) # the number of TALEs
#> [1] 11