Run the "build" stage of AnnoTALE

An R wrapper around the AnnoTALE 'AnnoTALE.jar build' program. It usually takes its input from the file generated by the run_annotale_predict function.

Usage

run_annotale_build(
  fasta_file,
  output_dir = getwd(),
  annotale_jar = system.file("tools", "AnnoTALEcli-1.5.jar", package = "tantale",
    mustWork = T)
)

Arguments

fasta_file

Path to a TALE_DNA_sequences_* fasta file, as written by run_annotale_predict's "predict" stage, to be classified into groups.

output_dir

Directory where output will be written (created if it does not exist).

annotale_jar

Path to the AnnoTALE jar file if you want to use another version than the one provided with tantale.

Value

0, invisibly; called for the files it writes to output_dir. If AnnoTALE exits with a non-zero status, the function stops with an error of class tantale_error_annotale_failed.

See also

Examples

# \donttest{
# Needs a Java runtime. Chained from run_annotale_predict()'s own output,
# since that is the file build normally classifies.
fasta <- system.file("extdata", "MAI1.fa", package = "tantale")
predict_out <- file.path(tempdir(), "annotale_build_example", "predict")
run_annotale_predict(fasta, output_dir = predict_out)
#> Running AnnoTALE predict for "MAI1"
#>   java -jar
#>   '/home/cunnac/Lab-Related/MyScripts/tantale/inst/tools/AnnoTALEcli-1.5.jar'
#>   predict g='/home/cunnac/Lab-Related/MyScripts/tantale/inst/extdata/MAI1.fa'
#>   s='MAI1' outdir='/tmp/RtmpqIoM6J/annotale_build_example/predict/Predict'
#> Running AnnoTALE analyze for "MAI1"
#>   java -jar
#>   /home/cunnac/Lab-Related/MyScripts/tantale/inst/tools/AnnoTALEcli-1.5.jar
#>   analyze
#>   t='/tmp/RtmpqIoM6J/annotale_build_example/predict/Predict/TALE_DNA_sequences_(MAI1).fasta'
#>   outdir='/tmp/RtmpqIoM6J/annotale_build_example/predict/Analyze'
predicted <- list.files(file.path(predict_out, "Predict"),
                        pattern = "^TALE_DNA_sequences_", full.names = TRUE)
build_out <- file.path(tempdir(), "annotale_build_example", "build")
run_annotale_build(predicted, output_dir = build_out)
#> Running AnnoTALE build
#>   java -Xms512M -Xmx6G -jar
#>   '/home/cunnac/Lab-Related/MyScripts/tantale/inst/tools/AnnoTALEcli-1.5.jar'
#>   build
#>   t='/tmp/RtmpqIoM6J/annotale_build_example/predict/Predict/TALE_DNA_sequences_(MAI1).fasta'
#>   outdir='/tmp/RtmpqIoM6J/annotale_build_example/build'
list.files(build_out)
#>  [1] "Class_1"             "Class_2"             "Class_3"            
#>  [4] "Class_4"             "Class_5"             "Class_6"            
#>  [7] "Class_7"             "Class_8"             "Class_9"            
#> [10] "Class_builder.xml"   "Tree_of_classes.pdf" "protocol_build.txt" 
# }